requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.6                # via pip-tools
dj-database-url==0.4.1
Django==1.10.2
djangorestframework==3.4.7
first==2.0.1              # via pip-tools
pip-tools==1.7.0
psycopg2==2.6.2
python-decouple==3.0
six==1.10.0               # via pip-tools
